+++ borax/utils.py
@@ -1,8 +1,7 @@
 # coding=utf8
 
-import re
-from functools import reduce
 import collections
+from functools import reduce
 
 
 def _resolve_value(val, args=None, kwargs=None):
@@ -93,11 +92,3 @@ def flatten(iterable):
             yield el
 
 
-def camel2snake(s):
-    camel_to_snake_regex = r'((?<=[a-z0-9])[A-Z]|(?!^)(?<!_)[A-Z](?=[a-z]))'
-    return re.sub(camel_to_snake_regex, r'_\1', s).lower()
-
-
-def snake2camel(s):
-    snake_to_camel_regex = r"(?:^|_)(.)"
-    return re.sub(snake_to_camel_regex, lambda m: m.group(1).upper(), s)